refactor(sketching): split draw.ts into drawing/pen/factories
Layer 3 audit F4 — `draw.ts` (542 LOC) mixed three concerns:

1. The immutable `Drawing` wrapper class
2. The mutable `DrawingPen` builder class (extends BaseSketcher2d)
3. 11 standalone factory functions (`draw`, `drawRectangle`,
   `drawCircle`, `drawEllipse`, `drawText`, etc.)

Split into three focused files:

- `drawing.ts` — `Drawing` class + `deserializeDrawing`
- `drawingPen.ts` — `DrawingPen` class + `draw()` factory
- `drawingFactories.ts` — 10 canned drawing factory functions

Public exports preserved via updated re-exports in `src/index.ts`,
`src/sketching.ts`, and tests.
