+++ src/sketching/drawing.ts
@@ -1,23 +1,11 @@
-import { unwrap } from '@/core/result.js';
 import { bug } from '@/core/errors.js';
-import type { ApproximationOptions, Curve2D } from '@/2d/lib/index.js';
-import {
-  BoundingBox2d,
-  deserializeCurve2D,
-  make2dCircle,
-  make2dEllipse,
-  make2dInerpolatedBSplineCurve,
-  make2dSegmentCurve,
-  type Point2D,
-  samePoint,
-} from '@/2d/lib/index.js';
+import type { ApproximationOptions } from '@/2d/lib/index.js';
+import { BoundingBox2d, deserializeCurve2D, type Point2D } from '@/2d/lib/index.js';
 import {
   Blueprint,
   cut2D,
   intersect2D,
   fuse2D,
-  polysidesBlueprint,
-  roundedRectangleBlueprint,
   type ScaleMode,
   type Shape2D,
   Blueprints,
@@ -26,17 +14,13 @@ import {
 import type { Plane, PlaneName } from '@/core/planeTypes.js';
 import type { PointInput } from '@/core/types.js';
 import type { AnyShape, Dimension, Face } from '@/core/shapeTypes.js';
-import { BaseSketcher2d } from '@/2d/blueprints/baseSketcher2d.js';
 import type { SketchInterface } from './sketch.js';
 import Sketches from './sketches.js';
-import type { GenericSketcher } from '@/2d/blueprints/genericSketcher.js';
 import type { SketchData } from '@/2d/blueprints/lib.js';
-import { textBlueprints } from '@/text/textBlueprints.js';
 
 import offsetFn, { type Offset2DConfig } from '@/2d/blueprints/blueprintOffset.js';
 import { cornerFinder, type CornerFinderFn } from '@/query/finderFns.js';
 import { fillet2D, chamfer2D } from '@/2d/blueprints/blueprintCustomCorners.js';
-import type { BSplineApproximationOptions } from '@/topology/shapeHelpers.js';
 import { approximateForSVG } from '@/2d/blueprints/blueprintApproximations.js';
 import type { SingleFace } from '@/query/helpers.js';
 import { wrapSketchData, wrapSketchDataArray } from './sketchFns.js';
@@ -303,57 +287,6 @@ export class Drawing {
 }
 
 /**
- * DrawingPen is a helper class to draw in 2D. It is used to create drawings
- * by exposing a builder interface. It is not a drawing itself, but it can be
- * used to create a drawing.
- *
- * @category Drawing
- */
-export class DrawingPen extends BaseSketcher2d implements GenericSketcher<Drawing> {
-  constructor(origin: Point2D = [0, 0]) {
-    super();
-    this.pointer = origin;
-    this.firstPoint = origin;
-
-    this.pendingCurves = [];
-  }
-
-  /** Finish drawing and return the resulting {@link Drawing} (does not close the path). */
-  done(): Drawing {
-    return new Drawing(new Blueprint(this.pendingCurves));
-  }
-
-  /** Close the path with a straight line to the start point and return the Drawing. */
-  close(): Drawing {
-    this._closeSketch();
-    return this.done();
-  }
-
-  /** Close the path by mirroring all curves about the line from first to last point. */
-  closeWithMirror(): Drawing {
-    this._closeWithMirror();
-    return this.close();
-  }
-
-  /**
-   * Close the path and apply a custom corner treatment between the last and first segments.
-   *
-   * @param radius - Fillet/chamfer radius.
-   * @param mode - Corner treatment type.
-   * @returns The closed {@link Drawing}.
-   */
-  closeWithCustomCorner(
-    radius: number | ((f: Curve2D, s: Curve2D) => Curve2D[]),
-    mode: 'fillet' | 'chamfer' | 'dogbone' = 'fillet'
-  ): Drawing {
-    this._closeSketch();
-    this._customCornerLastWithFirst(radius, mode);
-
-    return this.done();
-  }
-}
-
-/**
  * Deserializes a drawing from a string. String is expected to be in the format
  * generated by `Drawing.serialize()`.
  */
@@ -378,165 +311,3 @@ export function deserializeDrawing(data: string): Drawing {
   const shape = deserializeHelper(json);
   return new Drawing(shape);
 }
-
-/**
- * Creates a drawing pen to programatically draw in 2D.
- *
- * @category Drawing
- */
-export function draw(initialPoint?: Point2D): DrawingPen {
-  const pen = new DrawingPen();
-  if (initialPoint) {
-    pen.movePointerTo(initialPoint);
-  }
-  return pen;
-}
-
-/**
- * Creates the `Drawing` of a rectangle with (optional) rounded corners.
- *
- * The rectangle is centered on [0, 0]
- *
- * @category Drawing
- */
-export function drawRoundedRectangle(
-  width: number,
-  height: number,
-  r: number | { rx?: number; ry?: number } = 0
-): Drawing {
-  return new Drawing(roundedRectangleBlueprint(width, height, r));
-}
-/** Alias for {@link drawRoundedRectangle}. Creates a rectangle (sharp corners when `r` is 0). */
-export const drawRectangle = drawRoundedRectangle;
-
-/**
- * Creates the `Drawing` of a circle as one single curve.
- *
- * The circle is centered on [0, 0]
- *
- * @category Drawing
- */
-export function drawSingleCircle(radius: number): Drawing {
-  return new Drawing(new Blueprint([make2dCircle(radius)]));
-}
-
-/**
- * Creates the `Drawing` of an ellipse as one single curve.
- *
- * The ellipse is centered on [0, 0], with axes aligned with the coordinates.
- *
- * @category Drawing
- */
-export function drawSingleEllipse(majorRadius: number, minorRadius: number): Drawing {
-  const [minor, major] = [majorRadius, minorRadius].sort((a, b) => a - b) as [number, number];
-  const direction: Point2D = major === majorRadius ? [1, 0] : [0, 1];
-
-  return new Drawing(new Blueprint([make2dEllipse(major, minor, direction)]));
-}
-
-/**
- * Creates the `Drawing` of a circle.
- *
- * The circle is centered on [0, 0]
- *
- * @category Drawing
- */
-export function drawCircle(radius: number): Drawing {
-  return draw()
-    .movePointerTo([-radius, 0])
-    .sagittaArc(2 * radius, 0, radius)
-    .sagittaArc(-2 * radius, 0, radius)
-    .close();
-}
-
-/**
- * Creates the `Drawing` of an ellipse.
- *
- * The ellipse is centered on [0, 0], with axes aligned with the coordinates.
- *
- * @category Drawing
- */
-export function drawEllipse(majorRadius: number, minorRadius: number): Drawing {
-  return draw()
-    .movePointerTo([-majorRadius, 0])
-    .halfEllipse(2 * majorRadius, 0, minorRadius)
-    .halfEllipse(-2 * majorRadius, 0, minorRadius)
-    .close();
-}
-
-/**
- * Creates the `Drawing` of a polygon in a defined plane
- *
- * The sides of the polygon can be arcs of circle with a defined sagitta.
- * The radius defines the outer radius of the polygon without sagitta
- *
- * @category Drawing
- */
-export function drawPolysides(radius: number, sidesCount: number, sagitta = 0): Drawing {
-  return new Drawing(polysidesBlueprint(radius, sidesCount, sagitta));
-}
-
-/**
- * Creates the `Drawing` of a text, in a defined font size and a font family
- * (which will be the default).
- *
- * @category Drawing
- */
-export function drawText(
-  text: string,
-  { startX = 0, startY = 0, fontSize = 16, fontFamily = 'default' } = {}
-): Drawing {
-  return new Drawing(textBlueprints(text, { startX, startY, fontSize, fontFamily }));
-}
-
-/**
- * Creates the `Drawing` by interpolating points as a curve
- *
- * The drawing will be a spline approximating the points. Note that the
- * degree should be at maximum 3 if you need to export the drawing as an SVG.
- *
- * @category Drawing
- */
-export const drawPointsInterpolation = (
-  points: Point2D[],
-  approximationConfig: BSplineApproximationOptions = {},
-  options: {
-    closeShape?: boolean;
-  } = {}
-): Drawing => {
-  if (points.length < 2) {
-    bug(
-      'drawPointsInterpolation',
-      `Need at least 2 points for interpolation, got ${points.length}`
-    );
-  }
-  const curves = [unwrap(make2dInerpolatedBSplineCurve(points, approximationConfig))];
-  const firstPoint = points[0];
-  const lastPoint = points[points.length - 1];
-  if (options.closeShape && firstPoint && lastPoint && !samePoint(firstPoint, lastPoint)) {
-    curves.push(make2dSegmentCurve(lastPoint, firstPoint));
-  }
-
-  return new Drawing(new Blueprint(curves));
-};
-
-/**
- * Creates the `Drawing` of parametric function
- *
- * The drawing will be a spline approximating the function. Note that the
- * degree should be at maximum 3 if you need to export the drawing as an SVG.
- *
- * @category Drawing
- */
-export const drawParametricFunction = (
-  func: (t: number) => Point2D,
-  { pointsCount = 400, start = 0, stop = 1, closeShape = false } = {},
-  approximationConfig: BSplineApproximationOptions = {}
-): Drawing => {
-  const stepSize = (stop - start) / pointsCount;
-  const points = [...Array(pointsCount + 1).keys()].map((t) => {
-    return func(start + t * stepSize);
-  });
-
-  return drawPointsInterpolation(points, approximationConfig, { closeShape });
-};
