+++ tests/sketcher2d.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, beforeAll } from 'vitest';
 import { initKernel } from './setup.js';
 import FaceSketcher from '@/sketching/faceSketcher.js';
 import { BlueprintSketcher } from '@/2d/blueprints/blueprintSketcher.js';
-import { Drawing } from '@/sketching/draw.js';
+import { Drawing } from '@/sketching/drawing.js';
 import {
   draw,
   drawRectangle,
